common

+++ application/common/file.go
@@ -1,6 +1,7 @@
 package common
 
 import (
+	"github.com/krilie/lico_alone/common/common_struct/errs"
 	"github.com/krilie/lico_alone/common/context_util"
 	"github.com/krilie/lico_alone/module/file/service"
 	user2 "github.com/krilie/lico_alone/module/userbase/user"
@@ -24,6 +25,13 @@ func (AppCommon) DeleteFile(ctx *context_util.Context, userId, filePath string) 
 	_, e := user.GetInfo(ctx, userId)
 	if e != nil {
 		return e
+	}
+	file, e := fileOp.QueryByFilePath(ctx, userId, filePath)
+	if e != nil {
+		return e
+	}
+	if file.UserId != userId {
+		return errs.ErrNoPermission.NewWithMsg("can not find file you gaved")
 	}
 	return fileOp.DeleteFile(ctx, userId, filePath)
 }
